+++ lumicks/pylake/population/dwelltime.py
@@ -4,6 +4,8 @@ from scipy.optimize import minimize
 from dataclasses import dataclass, field
 import matplotlib.pyplot as plt
 from deprecated.sphinx import deprecated
+from lumicks.pylake.fitting.parameters import Parameter, Params
+from lumicks.pylake.fitting.profile_likelihood import ProfileLikelihood1D
 
 
 @dataclass(frozen=True)
@@ -361,6 +363,91 @@ class DwelltimeModel:
         # TODO: remove with deprecation
         self._bootstrap = bootstrap
         return bootstrap
+
+    def profile_likelihood(
+        self, *, num_steps=150, min_chi2_step=0.01, max_chi2_step=0.3, verbose=False
+    ):
+        """Calculate a likelihood profile.
+
+        This method traces an optimal path through parameter space in order to estimate parameter
+        confidence intervals. It iteratively performs a step for the profiled parameter, then
+        fixes that parameter and re-optimizes all the other parameters [1]_ [2]_.
+
+        Parameters
+        ----------
+        num_steps: integer
+            Number of steps to take .
+        min_chi2_step: float
+            Minimal desired step in terms of chi squared change prior to re-optimization. When the
+            step results in a fit change smaller than this threshold, the step-size will be
+            increased.
+        max_chi2_step: float
+            Minimal desired step in terms of chi squared change prior to re-optimization. When the
+            step results in a fit change bigger than this threshold, the step-size will be reduced.
+        verbose: bool
+            Controls the verbosity of the output.
+
+        References
+        ----------
+        .. [1] Raue, A., Kreutz, C., Maiwald, T., Bachmann, J., Schilling, M., Klingmüller, U.,
+               & Timmer, J. (2009). Structural and practical identifiability analysis of partially
+               observed dynamical models by exploiting the profile likelihood. Bioinformatics,
+               25(15), 1923-1929.
+        .. [2] Maiwald, T., Hass, H., Steiert, B., Vanlier, J., Engesser, R., Raue, A., Kipkeew,
+               F., Bock, H.H., Kaschek, D., Kreutz, C. and Timmer, J., 2016. Driving the model to
+               its limit: profile likelihood based model reduction. PloS one, 11(9).
+        """
+
+        def fit_func(params, lb, ub, fitted):
+            """Lower bound and upper bound are handled by _exponential_mle_optimize itself"""
+            optimized_params, _ = _exponential_mle_optimize(
+                self.n_components,
+                self.dwelltimes,
+                min_observation_time=self._observation_limits[0],
+                max_observation_time=self._observation_limits[1],
+                initial_guess=params,
+                options=self._optim_options,
+                fixed_param_mask=np.logical_not(fitted),
+            )
+
+            return optimized_params
+
+        def trial(params):
+            # Log-likelihood
+            return exponential_mixture_log_likelihood(
+                params,
+                self.dwelltimes,
+                min_observation_time=self._observation_limits[0],
+                max_observation_time=self._observation_limits[1],
+            )
+
+        # Pack parameters
+        keys = [f"amplitude {idx}" for idx in range(self.n_components)] + [
+            f"lifetime {idx}" for idx in range(self.n_components)
+        ]
+        lower_bounds = np.zeros(2 * self.n_components)
+        upper_bounds = np.hstack((np.ones(self.n_components), np.inf * np.ones(self.n_components)))
+        parameters = Params(
+            **{
+                key: Parameter(param, lower_bound=lb, upper_bound=ub)
+                for key, param, lb, ub in zip(keys, self._parameters, lower_bounds, upper_bounds)
+            }
+        )
+
+        def calculate_profile(param):
+            profile = ProfileLikelihood1D(
+                param,
+                num_dof=1,
+                min_chi2_step=min_chi2_step,
+                max_chi2_step=max_chi2_step,
+            )
+            profile._extend_profile(trial, fit_func, parameters, num_steps, True, verbose)
+            profile._extend_profile(trial, fit_func, parameters, num_steps, False, verbose)
+            return profile
+
+        return [
+            calculate_profile(param) for param in ([keys[-1]] if self.n_components == 1 else keys)
+        ]
 
     def pdf(self, x):
         """Probability Distribution Function (states as rows).